requirements.txt
psycopg2
cryptography
mysql-connector-python
ttkbootstrap
matplotlib
numpy